Added simulator (able to process 989/1000) tags

+++ InventoryManagement.Core/Entities/Product.cs
@@ -3,7 +3,7 @@
     public class Product : Entity
     {
         public string Name { get; set; }
-        public int ItemReference { get; set; }
+        public long ItemReference { get; set; }
 
         public int CompanyId { get; set; }
         public Company Company { get; set; }
