+++ InventoryManagement.Core/Repositories/ICompaniesRepository.cs
@@ -1,8 +1,9 @@
 ﻿using InventoryManagement.Core.Entities;
+
 namespace InventoryManagement.Core.Repositories
 {
     public interface ICompaniesRepository : IRepository<Company>
     {
-        Task<Company> GetCompanyByPrefixAsync(int companyPrefix);
+        Task<Company> GetCompanyByPrefixAsync(long companyPrefix);
     }
 }
