+++ InventoryManagement.Core/Repositories/IProductsRepository.cs
@@ -4,7 +4,7 @@ namespace InventoryManagement.Core.Repositories
 {
     public interface IProductsRepository : IRepository<Product>
     {
-        public Task<Product> GetProductByCompanyAndItemReferenceAsync(int companyId, int itemReference);
+        public Task<Product> GetProductByCompanyAndItemReferenceAsync(int companyId, long itemReference);
         public IDictionary<Company, int> GetProductsCountPerCompany();
         public IDictionary<Product, int> GetProductsCountPerProductByInventoryId(int inventoryId);
         public IDictionary<DateTime, IDictionary<Product, int>> GetProductsCountPerDayPerProduct();
